tests(chat): update tests after refactoring chat logic
Refs #156
